permisos usuarios

+++ app/Http/Controllers/Dashboard/PermissionController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Permission\PutRequest;
 use App\Http\Requests\Permission\StoreRequest;
 use Spatie\Permission\Models\Permission;
+use Illuminate\Support\Facades\Gate;
 
 class PermissionController extends Controller
 {
@@ -14,6 +15,7 @@ class PermissionController extends Controller
      */
     public function index()
     {
+          Gate::authorize('is-admin');
         $permissions = Permission::paginate(10);
         return view('dashboard/permission/index', compact('permissions'));
     }
@@ -23,6 +25,7 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        Gate::authorize('is-admin');
         $permission = new Permission();
         return view('dashboard.permission.create', compact('permission'));
     }
@@ -32,6 +35,7 @@ class PermissionController extends Controller
      */
     public function store(StoreRequest $request)
     {
+        Gate::authorize('is-admin');
         Permission::create($request->validated());
         return to_route('permission.index')->with('status', 'Permission created');
     }
@@ -41,6 +45,7 @@ class PermissionController extends Controller
      */
     public function show(Permission $permission)
     {
+        Gate::authorize('is-admin');
         return view('dashboard/permission/show', ['permission' => $permission]);
     }
 
@@ -49,6 +54,7 @@ class PermissionController extends Controller
      */
     public function edit(Permission $permission)
     {
+        Gate::authorize('is-admin');
         return view('dashboard.permission.edit', compact('permission'));
     }
 
@@ -57,6 +63,7 @@ class PermissionController extends Controller
      */
     public function update(PutRequest $request, Permission $permission)
     {
+        Gate::authorize('is-admin');
         $permission->update($request->validated());
         return to_route('permission.index')->with('status', 'Permission updated');
     }
@@ -66,6 +73,7 @@ class PermissionController extends Controller
      */
     public function destroy(Permission $permission)
     {
+        Gate::authorize('is-admin');
         $permission->delete();
         return to_route('permission.index')->with('status', 'Permission deleted');
     }
